dotCMS/core#17917 Implementing a custom ES client using HTTP 2
