+++ dotCMS/dependencies.gradle
@@ -1,10 +1,8 @@
 //noinspection GroovyAssignabilityCheck
 dependencies {
 
-    //compile group: 'com.dotcms.enterprise', name: 'ee', version: '5.0.0-SNAPSHOT', changing: true
+    compile group: 'com.dotcms.enterprise', name: 'ee', version: '5.0.0-SNAPSHOT', changing: true
 
-    compile fileTree(dir: 'src/main/enterprise/build/libs', include: ['ee_clean.jar'])
-    
     compile group: 'com.dotcms', name: 'ant-tooling', version: '1.3.2'
     compile ('io.jsonwebtoken:jjwt:0.6.0'){
         transitive = false
@@ -491,6 +489,8 @@ dependencies {
         exclude(group: 'org.apache.logging.log4j')
         exclude(group: 'com.fasterxml.jackson.core')
     }
+
+    implementation("com.squareup.okhttp3:okhttp:4.4.0")
 
     //https://www.oracle.com/database/technologies/appdev/jdbc-downloads.html
     providedCompile (group: 'com.oracle.jdbc', name: 'ojdbc8', version: '19.3') {
