--- dotCMS/src/main/java/com/dotcms/content/elasticsearch/util/HttpGetWithEntity.java
@@ -1,14 +1,0 @@
-package com.dotcms.content.elasticsearch.util;
-
-import org.apache.http.client.methods.HttpEntityEnclosingRequestBase;
-
-public class HttpGetWithEntity extends HttpEntityEnclosingRequestBase {
-
-    public final static String METHOD_NAME = "GET";
-
-    @Override
-    public String getMethod() {
-        return METHOD_NAME;
-    }
-
-}
